+++ examples/show_hidden_files.rb
@@ -8,9 +8,8 @@ require 'recls'
 #
 # 1. Ensure that they are returned in search, by including Recls::SHOW_HIDDEN
 # 2. Filter returned entries by hidden? attribute
-Recls::FileSearch.new('.', Recls::WILDCARDS_ALL, Recls::FILES | Recls::RECURSIVE | Recls::SHOW_HIDDEN).each do |fe|
+Recls.file_rsearch('.', Recls::WILDCARDS_ALL, Recls::FILES | Recls::SHOW_HIDDEN).each do |fe|
 
 	puts fe.path if fe.hidden?
-
 end