+++ CHANGES.md
@@ -2,7 +2,10 @@
 
 ## 2.10.1 - 14th April 2019
 
+* ~ fixed (Windows-only) defect whereby searching for both DIRECTORIES and FILES caused nothing to be found
 * + added examples/find_files_and_directories.rb
+* + added examples/find_files_and_directories.recursive.rb
+* + added explanatory .md files for all examples
 * ~ making examples executable
 * ~ hiding private/impl constructs from YARD